fix(predict): remove unnecessary casting SUITEDEV-35556

+++ android/src/main/kotlin/com/emarsys/emarsys_sdk/mapper/RecommendationLogicMapper.kt
@@ -73,8 +73,8 @@ class RecommendationLogicMapper : Mapper<Map<String, Any?>, Logic> {
         return cartItems.map {
             PredictCartItem(
                 it["itemId"] as String,
-                (it["price"] as Int).toDouble(),
-                (it["quantity"] as Int).toDouble()
+                it["price"] as Double,
+                it["quantity"] as Double
             )
         }.toList()
     }
